Before re-styling to black
