+++ src/TeamsComponent.tsx
@@ -2,6 +2,7 @@ import * as React from "react";
 
 import AFX from "./data";
 import { SingleTeamComponent } from "./SingleTeamComponent";
+import { DisplayUtility } from './DisplayUtility';
 
 //
 export class TeamsComponent extends React.Component<any> {
@@ -13,6 +14,18 @@ export class TeamsComponent extends React.Component<any> {
       let teamKey: string = teamIds[team];
       teamComponents.push(<SingleTeamComponent team={AFX.Groups[teamKey]} />);
     }
-    return <div className="teams">{teamComponents}</div>;
+
+    // Provide a date to label this group if teams for this semester
+    let dateStr: string = "";
+    if (teamIds.length > 0) {
+      dateStr = DisplayUtility.getSemesterString(AFX.Groups[teamIds[0]]);
+    }
+    return (
+      <div>
+        <div className="semester--title">{dateStr}</div>
+        <h2 className="board-team--title">Teams</h2>
+        <div className="teams">{teamComponents}</div>
+      </div>
+    );
   }
 }
